update ner module
